requirements.txt
python-dotenv
httpx
pydantic
streamlit
